refactor(server): rebalance handlers + adapter/identity boundaries into domain folders
Relocate 3 RPC handlers + 2 type/impl surfaces to their concept-owning
folders, fixing the adapters→app reverse-layer back-edge.

- presence handler (+ test): task/handlers/ → network/handlers/
- contacts handler: task/handlers/ → identity/handlers/
- connect handler: task/handlers/ → identity/handlers/ (auth handshake
  is an identity concern; fix the conversation-service relative import)
- ContactService interface: app/app-host.ts → identity/services/contact-policy.ts
  (app-host keeps the field/setter/getter, imports the type). Adapters now
  reach into identity (a lower layer) instead of back into app.
- WebhookSessionValidator impl: identity/services/session-validator.ts →
  adapters/webhook-session-validator.ts. The SessionValidator interface +
  SessionValidation type stay in identity (the contract).

Repoint every importer (app/server.ts handler barrel, app/types.ts,
standalone.ts, webhook adapters), the layer-tags.ts doc citations, and the
docs constants-snippet generators (generate-constants-snippets.ts,
generate-cli-docs.ts, check-no-hardcoded-constants.ts) that read HELLO
policy numbers from connect.handlers.ts; regenerate docs/snippets/constants.
Reconcile folder READMEs to #668 convention (no ARCHITECTURE.md).

Behavior-preserving: pure git-mv moves where possible. Build, eslint, knip,
oxlint, layer-typecheck, unit + integration tests all green.
