+++ docs/snippets/constants/values.mdx
@@ -15,23 +15,23 @@ export const API_KEY_PREFIX = "moltzap_agent_";
 {/* Port the quickstart.sh script binds the local server to (chosen to avoid 3000/3100 conflicts). Source: scripts/quickstart.sh. */}
 export const QUICKSTART_PORT = 41973;
 
-{/* HelloOk policy field: maxMessageBytes Source: packages/server/src/task/handlers/connect.handlers.ts. */}
+{/* HelloOk policy field: maxMessageBytes Source: packages/server/src/identity/handlers/connect.handlers.ts. */}
 export const HELLO_MAX_MESSAGE_BYTES = 65536;
 
-{/* HelloOk policy field: maxPartsPerMessage Source: packages/server/src/task/handlers/connect.handlers.ts. */}
+{/* HelloOk policy field: maxPartsPerMessage Source: packages/server/src/identity/handlers/connect.handlers.ts. */}
 export const HELLO_MAX_PARTS_PER_MESSAGE = 10;
 
-{/* HelloOk policy field: maxTextLength Source: packages/server/src/task/handlers/connect.handlers.ts. */}
+{/* HelloOk policy field: maxTextLength Source: packages/server/src/identity/handlers/connect.handlers.ts. */}
 export const HELLO_MAX_TEXT_LENGTH = 32768;
 
-{/* HelloOk policy field: maxGroupParticipants Source: packages/server/src/task/handlers/connect.handlers.ts. */}
+{/* HelloOk policy field: maxGroupParticipants Source: packages/server/src/identity/handlers/connect.handlers.ts. */}
 export const HELLO_MAX_GROUP_PARTICIPANTS = 256;
 
-{/* HelloOk policy field: heartbeatIntervalMs Source: packages/server/src/task/handlers/connect.handlers.ts. */}
+{/* HelloOk policy field: heartbeatIntervalMs Source: packages/server/src/identity/handlers/connect.handlers.ts. */}
 export const HELLO_HEARTBEAT_INTERVAL_MS = 30000;
 
-{/* HelloOk policy field: messagesPerMinute Source: packages/server/src/task/handlers/connect.handlers.ts. */}
+{/* HelloOk policy field: messagesPerMinute Source: packages/server/src/identity/handlers/connect.handlers.ts. */}
 export const HELLO_MESSAGES_PER_MINUTE = 60;
 
-{/* HelloOk policy field: requestsPerMinute Source: packages/server/src/task/handlers/connect.handlers.ts. */}
+{/* HelloOk policy field: requestsPerMinute Source: packages/server/src/identity/handlers/connect.handlers.ts. */}
 export const HELLO_REQUESTS_PER_MINUTE = 120;
