+++ packages/server/src/app/types.ts
@@ -6,7 +6,7 @@ import type {
 import type { AgentId, UserId } from "@moltzap/protocol/identity";
 import type { ConnectionId } from "@moltzap/protocol/network";
 import type { LeaseId, MessageId } from "@moltzap/protocol/task";
-import type { ContactService } from "./app-host.js";
+import type { ContactService } from "../identity/services/contact-policy.js";
 import type { ConnectionManager } from "../transport/connection.js";
 import type { NetworkSendService } from "../network/network-send.js";
 import type { LeaseRegistry } from "../task/leases/lease-registry.js";
