+++ packages/server/src/network/handlers/presence.handlers.test.ts
